+++ ArcdpsLogManager/CHANGELOG.md
@@ -14,6 +14,7 @@ This is the full changelog of the arcdps Log Manager.
 - Added AttackTarget event
 - Added SkillInfo event to SkillData
 - Added Last90BeforeDown event
+- Added Ruleset event
 - Added MissileCreate, MissileLaunch and MissileRemove events
 - Added EffectGroundCreate and EffectGroundRemove events
 - Added EffectAgentCreate and EffectAgentRemove events